+++ src/Symfony/Component/ObjectMapper/Metadata/ReverseClassObjectMapperMetadataFactory.php
@@ -62,8 +62,14 @@ final class ReverseClassObjectMapperMetadataFactory implements ObjectMapperMetad
             foreach ($attributes as $attribute) {
                 $map = $attribute->newInstance();
                 // We're forcing the target on a reverse mapping to the property name, doesn't make sense without a source
-                if ($map->source) {
-                    $mappings[] = new Mapping($reflProperty->getName(), $map->source, $map->if, $map->transform);
+                if ($map->sources || $map->source) {
+                    $mappings[] = new Mapping(
+                        $reflProperty->getName(),
+                        null,
+                        $map->sources ?: [$map->source],
+                        $map->if,
+                        $map->transform
+                    );
                 }
             }
         }