+++ src/Symfony/Component/ObjectMapper/Tests/Fixtures/Transform/TransformToStdClass.php
@@ -15,7 +15,7 @@ use Symfony\Component\ObjectMapper\TransformCallableInterface;
 
 class TransformToStdClass implements TransformCallableInterface
 {
-    public function __invoke(mixed $value, object $source, ?object $target): mixed
+    public function __invoke(mixed $value, object $source, ?object $target, mixed $targetValue): mixed
     {
         return new \stdClass();
     }