+++ src/Symfony/Component/ObjectMapper/Tests/Fixtures/TransformMultiSource/SubClassA.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Symfony\Component\ObjectMapper\Tests\Fixtures\TransformMultiSource;
+
+class SubClassA
+{
+    public function __construct(
+        public string $foo = 'foo'
+    ) {
+    }
+}