+++ src/Symfony/Component/ObjectMapper/Tests/Fixtures/TransformMultiSource/TargetClass.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Symfony\Component\ObjectMapper\Tests\Fixtures\TransformMultiSource;
+
+use Symfony\Component\ObjectMapper\Attribute\Map;
+
+class TargetClass
+{
+    public function __construct(
+        #[Map(sources: ['subClassA', 'subClassB'])]
+        public MultiplePropertiesClass $multiplePropertiesClass,
+    ) {
+    }
+}